+++ src/qdmi/devices/dd/Device.cpp
@@ -48,6 +48,7 @@
 #include <numeric>
 #include <ranges>
 #include <span>
+#include <sstream>
 #include <stdexcept>
 #include <string>
 #include <utility>
@@ -481,7 +482,6 @@ auto MQT_DDSIM_QDMI_Device_Job_impl_d::submitQIRProgram() -> QDMI_STATUS {
 auto MQT_DDSIM_QDMI_Device_Job_impl_d::submitQIRProgramSampling()
     -> QDMI_STATUS {
   return submitProgramAsync([this]() {
-    auto& runtime = qir::Runtime::getInstance();
     auto irBytes = std::visit(
         [](const auto& p) {
           return llvm::StringRef(reinterpret_cast<const char*>(p.data()),
@@ -489,12 +489,15 @@ auto MQT_DDSIM_QDMI_Device_Job_impl_d::submitQIRProgramSampling()
         },
         program_);
     auto jitSession = qir::JitSession(irBytes, "QDMI job");
+    auto& runtime = jitSession.runtime();
+    std::ostringstream output;
+    runtime.setOstream(output);
     runtime.outputProgramHeader();
     for (size_t i = 0; i < numShots_; ++i) {
       runtime.reset();
       runtime.outputShotStart();
       const auto rc = jitSession.run();
-      runtime.outputShotEnd();
+      runtime.outputShotEnd(rc);
       if (rc != 0) {
         throw std::runtime_error(
             llvm::formatv("QIR program failed with error: {}", rc));
@@ -515,16 +518,18 @@ auto MQT_DDSIM_QDMI_Device_Job_impl_d::submitQIRProgramStateExtraction()
     return QDMI_ERROR_NOTSUPPORTED;
   }
   return submitProgramAsync([this]() {
-    auto& runtime = qir::Runtime::getInstance();
-    runtime.reset();
     auto irBytes = std::visit(
         [](const auto& p) {
           return llvm::StringRef(reinterpret_cast<const char*>(p.data()),
                                  p.size());
         },
         program_);
-    auto jitSession =
-        qir::JitSession(irBytes, "QDMI job", qir::Execution::StateExtraction);
+    const qir::SessionOptions options{.execution =
+                                          qir::Execution::StateExtraction};
+    auto jitSession = qir::JitSession(irBytes, "QDMI job", options);
+    auto& runtime = jitSession.runtime();
+    std::ostringstream output;
+    runtime.setOstream(output);
     if (const auto rc = jitSession.run(); rc != 0) {
       throw std::runtime_error(
           llvm::formatv("QIR program failed with error: {}", rc));